Allow truncated device IDs to be specified in bdm.
Sort of like short git commit hashes, a device ID can be referred to using a
substring from the upper/least-significant side of the string if that substring
is unambiguous.

e.g. If a device OW0123456789AB exists, one could run either

    $ bdm console OW0123456789AB
    Connecting to OW0123456789AB

or

    $ bdm console 89AB
    Connecting to OW0123456789AB

with the same effect. On the other hand, if another device with id
OW0000000089AB exists, the command

    $ bdm console 89AB
    '89AB' is ambiguous. Be more specific.

will fail as ambiguous.

A new command, bdm expand <dev_id> will allow you to test the expansion of a
truncated device ID as it is performed/used in all commands that take device
identifiers.